requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mysqlclient==2.1.1
psycopg2==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
